+++ collect-webapp/frontend/src/datamanagement/components/recordeditor/fields/TaxonField/TaxonField.js
@@ -1,4 +1,4 @@
-import './TaxonField.css'
+import './TaxonField.scss'
 
 import React from 'react'