+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.8] - NOT RELEASED YET
+### Fixed
+- Hide not visible fields from taxon autocomplete;
+
 ## [4.0.7] - 2021-01-13
 ### Fixed
 - Issue with special characters in Portuguese